fixing my smelly pdf code
